feat: Added visit-store page
